mfapp/requirements.txt
fastapi
uvicorn
python-jose
requests
python-dotenv
pydantic
